# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Micro para gestión de usuarios, clientes y técnicos",
        "title": "Microservicio de usuario UDC",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/user-service/v1",
    "paths": {
...
        "/clients/{clientID}": {
            "get": {
                "description": "Busca un cliente por su UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Obtener cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del cliente",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Client"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/createClient": {
            "post": {
                "description": "Crea un nuevo cliente en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Crear cliente",
                "parameters": [
                    {
                        "description": "Datos del cliente",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Client"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/deleteClient/{clientID}": {
            "delete": {
                "description": "Elimina un cliente por su UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Eliminar cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del cliente",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejemplo: {\\\"message\\\":\\\"CLIENT_DELETED_SUCCESSFULLY\\\"}",
                        "schema": {
                            "$ref": "#/definitions/common.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/updateClient/{clientID}": {
            "put": {
                "description": "Actualiza los datos de un cliente existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Actualizar cliente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del cliente",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del cliente",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Client"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/common.ErrorResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/common.ErrorResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "common.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateClientRequest": {
            "type": "object",
            "required": [
                "Address",
                "Document",
                "DocumentTypeId",
                "Email",
                "Name",
                "PhoneNumber",
                "RoleId",
                "Surname"
            ],
            "properties": {
                "Address": {
                    "type": "string",
                    "example": "Calle 123 #45-67"
                },
                "Document": {
                    "type": "string",
                    "example": "1002003001"
                },
                "DocumentTypeId": {
                    "type": "string",
                    "example": "CC"
                },
                "Email": {
                    "type": "string",
                    "example": "juan.perez@example.com"
                },
                "Name": {
                    "type": "string",
                    "example": "Juan"
                },
                "PhoneNumber": {
                    "type": "string",
                    "example": "3001234567"
                },
                "RoleId": {
                    "type": "string",
                    "example": "CC"
                },
                "Surname": {
                    "type": "string",
                    "example": "Pérez"
                }
            }
        },
...
        "dto.UpdateClientRequest": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string",
                    "example": "Calle 123 #45-67"
                },
                "Document": {
                    "type": "string",
                    "example": "1002003001"
                },
                "DocumentTypeId": {
                    "type": "string",
                    "example": "CC"
                },
                "Email": {
                    "type": "string",
                    "example": "juan.perez@example.com"
                },
                "Name": {
                    "type": "string",
                    "example": "Juan"
                },
                "PhoneNumber": {
                    "type": "string",
                    "example": "3001234567"
                },
                "Surname": {
                    "type": "string",
                    "example": "Pérez"
                }
            }
        },
...
        "models.Client": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string",
                    "example": "Calle 123 #45-67"
                },
                "Document": {
                    "type": "string",
                    "example": "1002003001"
                },
                "DocumentType": {
                    "type": "string",
                    "example": "CC"
                },
                "Email": {
                    "type": "string",
                    "example": "juan.perez@example.com"
                },
                "ID": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "Name": {
                    "type": "string",
                    "example": "Juan"
                },
                "PhoneNumber": {
                    "type": "string",
                    "example": "3001234567"
                },
                "Surname": {
                    "type": "string",
                    "example": "Pérez"
                }
            }
        }
    }
}